Make call to std::pow unambigious
On OS X 10.11 (El Capitan) with LLVM / Clang the build fails due to
std::pow arguments being ambigious. This is because the std::pow
function is defined by using templates in LLVM's standard C++ library. This
isn't the case on all operating systems which is the reason why our CI
system didn't catch this.

Cast the first argument explicitly to double to make the function call
unambigious.

The CI system will be expanded to cover OS X in the near future to catch
situations like this.

+++ projects/lib/src/knockouttournament.cpp
@@ -216,8 +216,8 @@ QString KnockoutTournament::results() const
 					.arg(pair.first.score);
 			}
 			int r = round + 1;
-			int lineNum = (std::pow(2, r) - 1) +
-				      (x * std::pow(2, r + 1));
+			int lineNum = (std::pow((double)2, r) - 1) +
+				      (x * std::pow((double)2, r + 1));
 			QString text = QString(r * 2, '\t');
 			text += playerAt(winner).builder->name();
 			if (score != "0-0")
